requirements.txt
fastapi
uvicorn
pytest
sqlalchemy
alembic
psycopg2-binary
paho-mqtt
kafka-python
